+++ src/main/java/com/kolown/porring/account/repository/AccountRepository.java
@@ -1,0 +1,9 @@
+package com.kolown.porring.account.repository;
+
+import com.kolown.porring.account.entity.Account;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountRepository extends JpaRepository<Account, Long> {
+}
