+++ src/main/java/com/kolown/porring/account/repository/EmailAccountRepository.java
@@ -1,12 +1,15 @@
-package com.kolown.porring.security.repository;
+package com.kolown.porring.account.repository;
 
 import com.kolown.porring.account.entity.EmailAccount;
+
 import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface EmailAccountRepository extends JpaRepository<EmailAccount, Long> {
     boolean existsByEmail(String email);
+
     Optional<EmailAccount> findByEmail(String email);
 }
